+++ src/ModularPipelines/Attributes/NotInParallelAttribute.cs
@@ -1,0 +1,4 @@
+﻿namespace ModularPipelines.Attributes;
+
+[AttributeUsage(AttributeTargets.Class)]
+public class NotInParallelAttribute : Attribute;