+++ src/ModularPipelines/Engine/IModuleExecutor.cs
@@ -4,5 +4,6 @@ namespace ModularPipelines.Engine;
 
 internal interface IModuleExecutor
 {
-    Task<IEnumerable<ModuleBase>> ExecuteAsync(IEnumerable<ModuleBase> modules);
+    Task<IEnumerable<ModuleBase>> ExecuteAsync(IReadOnlyList<ModuleBase> modules);
+    Task<ModuleBase> ExecuteAsync(ModuleBase module);
 }