+++ src/ModularPipelines/Engine/ModuleExecutor.cs
@@ -1,4 +1,6 @@
+using System.Reflection;
 using Microsoft.Extensions.Options;
+using ModularPipelines.Attributes;
 using ModularPipelines.Extensions;
 using ModularPipelines.Modules;
 using ModularPipelines.Options;
@@ -24,20 +26,40 @@ internal class ModuleExecutor : IModuleExecutor
         _moduleDisposer = moduleDisposer;
     }
 
-    public async Task<IEnumerable<ModuleBase>> ExecuteAsync(IEnumerable<ModuleBase> modules)
+    public async Task<IEnumerable<ModuleBase>> ExecuteAsync(IReadOnlyList<ModuleBase> modules)
     {
-        var moduleTasks = modules.Select(ExecuteAsync).ToArray();
-
+        var moduleResults = new List<ModuleBase>();
+        
+        var nonParallelModules = modules
+            .Where(x => x.GetType().GetCustomAttribute<NotInParallelAttribute>() != null)
+            .ToList();
+        
+        foreach (var nonParallelModule in nonParallelModules.OrderBy(x => x.DependentModules.Count))
+        {
+            moduleResults.Add(await ExecuteAsync(nonParallelModule));
+        }
+        
+        var moduleTasks = modules.Except(nonParallelModules).Select(ExecuteAsync).ToArray();
+        
         if (_pipelineOptions.Value.ExecutionMode == ExecutionMode.StopOnFirstException)
         {
-            return await moduleTasks.WhenAllFailFast();
+            moduleResults.AddRange(await moduleTasks.WhenAllFailFast());
+        }
+        else
+        {
+            moduleResults.AddRange(await Task.WhenAll(moduleTasks));
         }
 
-        return await Task.WhenAll(moduleTasks);
+        return moduleResults;
     }
 
-    private async Task<ModuleBase> ExecuteAsync(ModuleBase module)
+    public async Task<ModuleBase> ExecuteAsync(ModuleBase module)
     {
+        if (module.IsStarted)
+        {
+            return module;
+        }
+        
         try
         {
             await _pipelineSetupExecutor.OnBeforeModuleStartAsync(module);