+++ src/ModularPipelines.Build/ReleaseNotes.md
@@ -1,1 +1,2 @@
 *   Chocolatey CLI helpers
+- Not in parallel attribute for running modules without parallelisation